Split encode.ts into codes/encode.ts and codecs/decode.ts
