+++ ethereum_history_api/oracles/src/noir/oracles/codec/const.ts
@@ -1,0 +1,7 @@
+export const MODULUS = 21888242871839275222246405745257275088696311157297823662689037894645226208583n;
+export const BYTES32_LENGTH = 32;
+export const ADDRESS_LENGTH = 20;
+export const BITS_IN_BYTE = 8n;
+export const MAX_U8 = 255;
+export const PROOF_ONE_LEVEL_LENGTH = 532;
+export const ZERO_PAD_VALUE = '0x0';
